Fix crash on Overlay DestroySwapchain
If a swapchain is currently in the overlay quad layers, then it is
expected to be passed off at Main session EndFrame.  So defer its
destruction until it is no longer in the quad layers at EndFrame.

Otherwise, a swapchain could be destroyed and then subsequently handed
to EndFrame, which is an error.