add strings for XR_EXT_hp_mixed_reality_controller
Microsoft's "latest preview OpenXR runtime" 104.2010.27002 includes the
".../hp/.." interaction profile.  This commit adds the WellKnownString
entries for that so that stand-in XrActions can be created on the main
side for that profile.

Note that 104.2010.27002 has a exception after startup
preventing it from running in the demo scene, may or may not be related
to the API layer.