Add envirnoment blendmode alpha to the overlay enum
Even if the underlying device *doesn't* support alpha blending, report
that it does, to allow overlay sessions to specify it in their endframe
submissions.

NOTE: This should be conditional for the case of a client instance, but
currently we don't know if the application will be a client until
BeginSession time.